fix(history): fixed regression issue which added double slashes to the pushState fragment
The previous commit which introduced a forward slash in-front of url fragments causes double up with child routes. Not sure if this should go deeper, but the easiest solution is to check if the first character is a forward slash and if it is, then don't add another. References issue [#141](https://github.com/aurelia/router/issues/141) which was created on the Router component.