fix(history-browser): Add location.href to replaceState() call
There is currently a bug in Safari that doesn't properly handle the optional third parameter to the replaceState(state, title, url) function when a <base /> tag is present on the page. This fix explicitly sets the url property to the location.href value.

Thanks to @powerbuoy @huochunpeng and @jwx for helping diagnose this issue!

Closes https://github.com/aurelia/history-browser/issues/34 https://github.com/aurelia/router/issues/574