fix(index): revert return types
The return types were updating to be more accurage, indicating the possible Promise return values. However, the return types in the history base class were not. As a result, the TS definitions will not build and this would render TS devs unable to use the module. I've reverted this for now. Later, we can first update history and the history-browser so that there's proper compatability. This would still be a breaking change for TS users but at least the code would compile.